Move the test inside of the common tests file
